Move akmaev convective adjusment code to its own module.
Increment version number to 0.6.2.dev0
