+++ climlab/convection/convadj.py
@@ -4,7 +4,7 @@ import numpy as np
 from climlab import constants as const
 from climlab.process.time_dependent_process import TimeDependentProcess
 from climlab.domain.field import Field
-import sys
+from .akmaev_adjustment import convective_adjustment_direct
 
 
 class ConvectiveAdjustment(TimeDependentProcess):
@@ -74,202 +74,3 @@ class ConvectiveAdjustment(TimeDependentProcess):
         #  return the adjustment, independent of timestep
         #  because the parent process might have set a different timestep!
         return self.adjustment
-
-
-def convective_adjustment_direct(p, T, c, lapserate=6.5):
-    """Convective Adjustment to a specified lapse rate.
-
-    Input argument lapserate gives the lapse rate expressed in degrees K per km
-    (positive means temperature increasing downward).
-
-    Default lapse rate is 6.5 K / km.
-
-    Returns the adjusted Column temperature.
-    inputs:
-    p is pressure in hPa
-    T is temperature in K
-    c is heat capacity in in J / m**2 / K
-    """
-    # largely follows notation and algorithm in Akamaev (1991) MWR
-
-    #if lapserate is 'DALR':
-    #    lapserate = const.g / const.cp * 1.E3
-    #try:
-    #    alpha = const.Rd / const.g * lapserate / 1.E3
-    #except:
-    #    raise ValueError('Problem with lapse rate')
-    #lapserate = lapserate * np.ones(T.shape)  # same dimensions as T
-    alpha = const.Rd / const.g * lapserate / 1.E3 # same dimensions as lapserate
-    L = p.size
-    #  Here const.ps = 1000 hPa is just used as a reference pressure
-    #  to compute potential temperature, so is valid even for different
-    #  surface pressures
-    Pi = (p[:]/const.ps)**alpha  # will need to modify to allow variable lapse rates
-    beta = 1./Pi
-    theta = T * beta
-    q = Pi * c
-    n_k = np.zeros(L, dtype=np.int8)
-    theta_k = np.zeros_like(p)
-    s_k = np.zeros_like(p)
-    t_k = np.zeros_like(p)
-    thetaadj = Akamaev_adjustment_multidim(theta, q, beta, n_k,
-                                           theta_k, s_k, t_k)
-    T = thetaadj * Pi
-    return T
-
-
-# @jit  # numba.jit not working here. Not clear why.
-#  At least we get something like 10x speedup from the inner loop
-def Akamaev_adjustment_multidim(theta, q, beta, n_k, theta_k, s_k, t_k):
-    L = q.size  # number of vertical levels
-    size0 = theta.shape[0] #np.size(T, axis=0)
-    if size0 != L:
-        num_lat = size0
-        for lat in range(num_lat):
-            theta[lat,:] = Akamaev_adjustment(theta[lat,:], q, beta, n_k,
-                                              theta_k, s_k, t_k)
-    else:
-        num_lat = 1
-        theta = Akamaev_adjustment(theta, q, beta, n_k, theta_k, s_k, t_k)
-    return theta
-
-
-def Akamaev_adjustment(theta, q, beta, n_k, theta_k, s_k, t_k):
-    '''Single column only.'''
-    L = q.size  # number of vertical levels
-    # Akamaev step 1
-    k = 1
-    n_k[k-1] = 1
-    theta_k[k-1] = theta[k-1]
-    l = 2
-    while True:
-        # Akamaev step 2
-        n = 1
-        thistheta = theta[l-1]
-        while True:
-            # Akamaev step 3
-            if theta_k[k-1] <= thistheta:
-                # Akamaev step 6
-                k += 1
-                break  # to step 7
-            else:
-                if n <= 1:
-                    s = q[l-1]
-                    t = s*thistheta
-                # Akamaev step 4
-                if n_k[k-1] <= 1:
-                # lower adjacent level is not an earlier-formed neutral layer
-                    s_k[k-1] = q[l-n-1]
-                    t_k[k-1] = s_k[k-1] * theta_k[k-1]
-                # Akamaev step 5
-                    #  join current and underlying layers
-                n += n_k[k-1]
-                s += s_k[k-1]
-                t += t_k[k-1]
-                s_k[k-1] = s
-                t_k[k-1] = t
-                thistheta = t/s
-                if k==1:
-                    #  joint neutral layer is the first one
-                    break  # to step 7
-                k -= 1
-                # back to step 3
-        # Akamaev step 7
-        if l == L:  # the scan is over
-            break  # to step 8
-        l += 1
-        n_k[k-1] = n
-        theta_k[k-1] = thistheta
-        # back to step 2
-
-
-#
-#    for l in range(2, L+1):  # l = 2,3,4,...,L
-#        n = 1
-#        thistheta = theta[l-1]
-#        done = False
-#        while not done:
-#            if (theta_k[k-1] > thistheta):
-#                s = 0.
-#                t = 0.
-#                # stratification is unstable
-#                if n == 1:
-#                    # current layer is not an earlier-formed neutral layer
-#                    s = q[l-1]
-#                    t = s * thistheta
-#                if (n_k[k-1] < 2):
-#                    # lower adjacent level is not an earlier-formed neutral layer
-#                    s_k[k-1] = q[l-n-1]
-#                    t_k[k-1] = s_k[k-1] * theta_k[k-1]
-#                #  join current and underlying layers
-#                n += n_k[k-1]
-#                s += s_k[k-1]
-#                s_k[k-1] = s
-#                t += t_k[k-1]
-#                t_k[k-1] = t
-#                thistheta = t / s
-#                if k == 1-1:
-#                    # joint neutral layer in the first one, done checking lower layers
-#                    done = True
-#                else:
-#                    k -= 1
-#                    # go back and check stability of the lower adjacent layer
-#            else:
-#                k += 1  # statification is stable
-#                done = True
-#        if l < L:
-#            n_k[k-1] = n
-#            theta_k[k-1] = thistheta
-#    #  finished looping through to check stability
-
-    # update the potential temperatures
-    while True:
-        while True:
-            # Akamaev step 8
-            if n==1: # current model level was not included in any neutral layer
-                break  # to step 11
-            while True:
-                #  Akamaev step 9
-                theta[l-1] = thistheta
-                if n==1:
-                    break
-                #  Akamaev step 10
-                l -= 1
-                n -= 1
-                #  back to step 9
-        # Akamaev step 11
-        if k==1:
-            break
-        k -= 1
-        l -= 1
-        n = n_k[k-1]
-        thistheta = theta_k[k-1]
-        # back to step 8
-
-#    for l in range(L, 0, -1):
-#        if n > 1:
-#        while n>1:
-#            theta[l-1] = thistheta
-#            n -= 1
-#        else:
-#            k -= 1
-#            n = n_k[k]
-#            thistheta = theta_k[k]
-    return theta
-
-#  Attempt to use numba to compile the Akamaev_adjustment function
-#  which gives at least 10x speedup
-#   If numba is not available or compilation fails, the code will be executed
-#   in pure Python. Results should be identical
-
-###  ONLY DO THIS IN PYTHON 2!
-#  Because of a bug in numba for while / break loops that is causing all calls
-#  to ConvectiveAdjustment to hang forever!
-#   https://github.com/numba/numba/issues/2273
-if sys.version_info < (3,0):
-    try:
-        from numba import jit
-        Akamaev_adjustment = jit(signature_or_function=Akamaev_adjustment)
-        #print 'Compiling Akamaev_adjustment() with numba.'
-    except:
-        pass
