+++ conda-recipe/meta.yaml
@@ -1,5 +1,5 @@
 {% set name = "climlab" %}
-{% set version = "0.6.1" %}
+{% set version = "0.6.2.dev0" %}
 
 package:
   name: {{ name|lower }}
